Widgets: Move SidebarPage to sidebardialog module
